requirements.txt
yfinance
streamlit
joblib
tensorflow-cpu
scikit-learn
pandas_market_calendars